refactor: 상태에 따라 컴포넌트 분리가 아닌 재사용하기 위한 컴포넌트 구현
